[frontend] split component to files (#9760)
